requirements.txt
TM1py>=2.0.0
anthropic>=0.40.0
openai>=1.0.0
voyageai>=0.3.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
tenacity>=9.0.0
PyYAML>=6.0
# Web search
ddgs>=9.0.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
# Google Drive (optional — install if GOOGLE_DRIVE_ENABLED=true)
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
